src/sb_consumer/requirements.txt
azure-identity
azure-servicebus
azure-storage-blob
python-dotenv
Faker